requirements.txt
aiofiles
aiogram
aiohttp[speedups]
aioredis
alphabet_detector
asyncio
beautifulsoup4==4.11.1
better_profanity
bing_image_downloader
bleach
bs4
cachetools
cloudscraper
countryinfo
cowpy
cryptg
cryptography
dateparser==1.1.1
dnspython
emoji==1.7.0
envparse
faker
feedparser
ffmpeg-python
fontTools
future
geopy
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
glitch_this
glitch_this
google-api-python-client
google-trans-new
google-trans-new
googletrans
gpytranslate
gtts
hachoir
heroku3
hiredis
html-telegraph-poster
html2text
httpx
humanize
img2pdf
jikanpy
launchpadlib==1.10.17
lxml
lyricsgenius
markdown
markdown2>=2.4.0
motor<2.4.0,>=2.1.0
multicolorcaptcha
nekos.py
nltk
nudepy
NumPy
odmantic
opencv-python-headless
pendulum
Pillow
pretty_errors
psutil
psycopg2-binary
PyDictionary==2.0.1
pydub
pykeyboard
pymongo>=3.10
pynewtonmath
pyrate-limiter
pyrogram==1.4.16
python-arq
python-dotenv
python-rapidjson
python-telegram-bot==13.14
pytimeparse
pytube
pytz
pyYAML>=5.1.2
redis
regex
requests
requests_html
search_engine_parser
secureme
selenium
spamwatch
speedtest-cli
spongemock
SQLAlchemy==1.4.36
telegraph
telethon==1.25.4
textblob
tgcrypto
tswift
ujson
uvloop
wget
wikipedia
youtube_dl
youtube_search
youtube_search_python
yt-dlp
zalgo-text
#ǫ ʀᴇ sᴀʟᴇ ɢᴀʏ ᴋʏ ᴅᴇᴋ ʀᴀ ʜ ʏᴀ 
